Rebase and correct code per code review PR

+++ api/src/main/java/org/openmrs/module/attachments/obs/ImageAttachmentHandler.java
@@ -2,15 +2,16 @@ package org.openmrs.module.attachments.obs;
 
 import java.io.File;
 import java.io.IOException;
-
+import java.awt.image.BufferedImage;
+import javax.imageio.ImageIO;
+import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.lang.StringUtils;
 import org.openmrs.Obs;
 import org.openmrs.api.APIException;
 import org.openmrs.module.attachments.AttachmentsConstants;
 import org.openmrs.obs.ComplexData;
 import org.openmrs.obs.handler.AbstractHandler;
 import org.openmrs.obs.handler.ImageHandler;
-
-import net.coobird.thumbnailator.Thumbnails;
 
 public class ImageAttachmentHandler extends AbstractAttachmentHandler {
 	
@@ -31,7 +32,8 @@ public class ImageAttachmentHandler extends AbstractAttachmentHandler {
 	protected ComplexData readComplexData(Obs obs, ValueComplex valueComplex, String view) {
 		
 		String fileName = valueComplex.getFileName();
-		if (view.equals(AttachmentsConstants.ATT_VIEW_THUMBNAIL)) {
+		if (view.equals(AttachmentsConstants.ATT_VIEW_THUMBNAIL)&& 
+		  !StringUtils.endsWith(FilenameUtils.removeExtension(fileName),NO_THUMBNAIL_SUFFIX)) {
 			fileName = buildThumbnailFileName(fileName);
 		}
 		
@@ -51,39 +53,55 @@ public class ImageAttachmentHandler extends AbstractAttachmentHandler {
 		
 		// We use a temp obs whose complex data points to the file names
 		String fileName = complexData.getTitle();
-		String thumbnailFileName = buildThumbnailFileName(fileName);
-		
+	    String thumbnailFileName = null;
+	    boolean isThumbNailPurged = false;
 		Obs tmpObs = new Obs();
-		tmpObs.setValueComplex(thumbnailFileName);
-		boolean isThumbNailPurged = getParent().purgeComplexData(tmpObs);
+
+	    if (!StringUtils.endsWith(FilenameUtils.removeExtension(fileName), NO_THUMBNAIL_SUFFIX)) {
+	    	thumbnailFileName = buildThumbnailFileName(fileName);
+			tmpObs.setValueComplex(thumbnailFileName);
+			isThumbNailPurged = getParent().purgeComplexData(tmpObs);
+	    }
+		
 		tmpObs.setValueComplex(fileName);
 		boolean isImagePurged = getParent().purgeComplexData(tmpObs);
 		
-		return isThumbNailPurged && isImagePurged;
+		return isThumbNailPurged || isImagePurged;
 	}
 	
 	@Override
 	protected ValueComplex saveComplexData(Obs obs, AttachmentComplexData complexData) {
-		
+		int imageHeight = Integer.MAX_VALUE;
+		int imageWidth = Integer.MAX_VALUE;
+		String savedFileName = null;
+	
 		// We invoke the parent to inherit from the file saving routines.
 		obs = getParent().saveObs(obs);
-		
+
 		File savedFile = AbstractHandler.getComplexDataFile(obs);
-		String savedFileName = savedFile.getName();
-		
-		// Saving the thumbnail
-		File dir = savedFile.getParentFile();
-		String thumbnailFileName = buildThumbnailFileName(savedFileName);
+		// String savedFileName = savedFile.getName();
+
+		// Get image dimensions
 		try {
-			Thumbnails.of(savedFile.getAbsolutePath()).size(THUMBNAIL_MAX_HEIGHT, THUMBNAIL_MAX_WIDTH)
-			        .toFile(new File(dir, thumbnailFileName));
-		}
-		catch (IOException e) {
+			BufferedImage image = ImageIO.read(savedFile);
+			imageHeight = image.getHeight();
+			imageWidth = image.getWidth();
+		} catch (IOException e) {
 			getParent().purgeComplexData(obs);
-			throw new APIException("A thumbnail file could not be saved for obs with" + "OBS_ID='" + obs.getObsId() + "', "
-			        + "FILE='" + complexData.getTitle() + "'.", e);
+		         throw new APIException("Can't read the image file"
+		               + "OBS_ID='" + obs.getObsId() + "', "
+		               + "FILE='" + complexData.getTitle() + "'.", e);
 		}
-		
+
+		try {
+			savedFileName = saveThumbnailOrRename(savedFile, imageHeight, imageWidth);
+		} catch (APIException e) {
+			getParent().purgeComplexData(obs);
+	        	throw new APIException("A thumbnail file could not be saved for obs with"
+	               + "OBS_ID='" + obs.getObsId() + "', "
+	               + "FILE='" + complexData.getTitle() + "'.", e);
+		}
+
 		return new ValueComplex(complexData.getInstructions(), complexData.getMimeType(), savedFileName);
 	}
 }
